Iterate org-level Renovate configuration
This PR tries to simplify the renovate config by using best practices
and builtin Renovate features.

This results in a simpler config file,
that should still obtain the same level of clarity and grouping in the
generated PRs.

I don't exclude that further iterations might happen, to further improve
or adjust the type of PRs being generated.

Ref: https://issues.redhat.com/browse/EC-1417